Swaps to std::atan2

+++ shared/utility/math/angle.hpp
@@ -167,7 +167,7 @@ namespace utility::math::angle {
 
         // Find alpha such that $\sin(\alpha) = a\_$ and $\cos(\alpha) = b\_$, which is possible because $a\_^2
         // + b\_^2 = 1$
-        float alpha = atan2f(a_, b_);
+        float alpha = std::atan2(a_, b_);
 
         // Hence the equation becomes $\cos(\alpha)\cos(x)+\sin(\alpha)\sin(x) = cos(x-\alpha) = c\_$
         return alpha + acos_clamped(c_);
